requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3 
requests==2.23.0
authlib
google-api-python-client==1.10.0
google-auth==1.19.2